+++ packages/histoire/src/node/build.ts
@@ -11,6 +11,7 @@ import { lookup as lookupMime } from 'mrmime'
 import pc from 'picocolors'
 import { performance } from 'node:perf_hooks'
 import type {
+  ChangeViteConfigCallback,
   BuildEndCallback,
   PreviewStoryCallback,
 } from '@histoire/shared'
@@ -54,12 +55,14 @@ export async function build (ctx: Context) {
     server,
     throws: true,
   })
+  const changeViteConfigCallbacks: ChangeViteConfigCallback[] = []
   const buildEndCallbacks: BuildEndCallback[] = []
   const previewStoryCallbacks: PreviewStoryCallback[] = []
   for (const plugin of ctx.config.plugins) {
     if (plugin.onBuild) {
       const api = new BuildPluginApi(ctx, plugin, moduleLoader)
       await plugin.onBuild(api)
+      changeViteConfigCallbacks.push(...api.changeViteConfigCallbacks)
       buildEndCallbacks.push(...api.buildEndCallbacks)
       previewStoryCallbacks.push(...api.previewStoryCallbacks)
     }
@@ -140,6 +143,16 @@ export async function build (ctx: Context) {
       config.build.rollupOptions.output = {
         manualChunks (id) {
           if (!id.includes('@histoire/app') && id.includes('node_modules')) {
+            for (const test of ctx.config.build?.excludeFromVendorsChunk ?? []) {
+              if ((
+                typeof test === 'string' && id.includes(test)
+              ) || (
+                test instanceof RegExp && test.test(id)
+              )) {
+                // Excluded from vendor chunk
+                return
+              }
+            }
             return 'vendor'
           }
         },
@@ -154,8 +167,15 @@ export async function build (ctx: Context) {
         // Don't build in SSR mode
         ssr: false,
       })
+
+      config.define.__HST_COLLECT__ = false
     },
   })
+
+  for (const cb of changeViteConfigCallbacks) {
+    console.log('vite config hook', cb)
+    await cb(buildViteConfig)
+  }
 
   const results = await viteBuild(buildViteConfig)
   const result = Array.isArray(results) ? results[0] : results as RollupOutput